Query for dimensions to apply to the screen when receiving an XRRCrtcChangeNotifyEvent
When receiving this event, we're told the dimensions of the CRTC. These
dimensions apparently do not reflect the actual dimensions of the CRTC.
This means ratpoison's knowledge of screen size can be inaccurate and
ratpoison will use only a portion of the screen.

This change makes us query for the dimensions of the CRTC and then use
those rather than use the dimensions provided by the event.

This addresses a problem with different sized screens after using the
xrandr command. In my setup I have two different sized screens with
different DPIs (a laptop and a monitor). Upon logging on, I issue an
xrandr command to scale the monitor's output. Prior to this change,
ratpoison used only part of the monitor after doing this (the top left).
Restarting ratpoison is a workaround to make it aware of the monitor's
dimensions, but an unsatisfactory one.

I believe the issue stems from scaling for mixed DPI. The CRTC change
event appears to have a width and height prior to applying scaling
(2x2) (1920x1080 vs. 3840x2160 in my case).